test(unit): converted the questions tests to vitest
