+++ .eslintrc.yml
@@ -1,3 +1,7 @@
 extends:
   - '@travi'
-  - '@travi/mocha'
+
+overrides:
+  - files: ./vitest.config.js
+    rules:
+      import/no-unresolved: off
