+++ package.json
@@ -15,7 +15,7 @@
     "test": "npm-run-all --print-label --parallel lint:* --parallel test:*",
     "lint:sensitive": "ban",
     "test:unit": "cross-env NODE_ENV=test c8 run-s test:unit:base",
-    "test:unit:base": "DEBUG=any mocha --recursive test/unit",
+    "test:unit:base": "DEBUG=any vitest run",
     "lint:js": "eslint . --cache",
     "lint:md": "remark . --frail",
     "clean": "rimraf lib/",
@@ -41,24 +41,23 @@
     "@travi/any": "3.1.2",
     "@travi/babel-preset": "3.0.131",
     "@travi/eslint-config": "5.0.10",
-    "@travi/eslint-config-mocha": "1.0.25",
     "ban-sensitive-files": "1.10.6",
     "c8": "10.1.2",
-    "chai": "4.5.0",
     "commitlint-config-travi": "1.4.50",
     "cross-env": "7.0.3",
     "cz-conventional-changelog": "3.3.0",
     "husky": "9.1.6",
     "inquirer": "8.2.6",
+    "jest-when": "3.6.0",
     "lockfile-lint": "4.14.0",
-    "mocha": "10.7.3",
     "npm-run-all2": "6.2.3",
     "publint": "0.2.11",
     "remark-cli": "12.0.1",
     "remark-preset-lint-travi": "6.0.5",
     "rimraf": "6.0.1",
     "rollup": "4.22.4",
-    "rollup-plugin-auto-external": "2.0.0"
+    "rollup-plugin-auto-external": "2.0.0",
+    "vitest": "2.1.1"
   },
   "peerDependencies": {
     "inquirer": "^6.0.0 || ^7.0.0 || ^8.0.0"
