--- test/mocha-setup.js
@@ -1,3 +1,0 @@
-process.on('unhandledRejection', reason => {
-  throw reason;
-});
