+++ vitest.config.js
@@ -1,0 +1,8 @@
+import {defineConfig} from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    restoreMocks: true
+  }
+});
